requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.7
Werkzeug==3.0.3
python-dotenv==1.0.1
reportlab==4.2.2
Pillow==10.4.0
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
WTForms==3.1.2
email-validator==2.2.0
